cmp: Avoid tripping cmp's cyclic transformer check
After I initially implemented this, upstream go-cmp introduced a (pretty
reasonable) recursion limit on transformers to detect when they are just
feeding their own results back into themselves.

Unfortunately, this package's cmp transformer now often ends up being
flagged false-positive for that check because we legitimately do have
nested transformations to make.

To mitigate that problem, we'll now do all of the recursion internally
inside the functions in this package, so that go-cmp believes it's
performing only a single layer of transformation and doesn't find itself
having to call the same transformer again to deal with nested values.