cmp: Don't panic if a marked value needs transforming
For the subset of types that we transform into simpler cmpable types, we
need to unmark the value before trying to disassemble it for comparison.

However, we don't want to erase the marks entirely since that would make
it impossible to test whether marking is happening correctly, so instead
we wrap any transformation of a marked value in a new struct type
ctyMarkedVal, which then allows the marks to remain associated with the
value even though it isn't a cty.Value anymore.